p1_CT-TGNN_code/requirements.txt
# Core Deep Learning
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
torch-geometric==2.3.1
torch-scatter==2.1.1
torch-sparse==0.6.17
torch-cluster==1.6.1
torch-spline-conv==1.2.2

# Neural ODE and Differential Equations
torchdiffeq==0.2.3

# Scientific Computing
numpy==1.24.3
scipy==1.10.1
pandas==2.0.2
scikit-learn==1.2.2

# Graph Processing
networkx==3.1
igraph==0.10.4

# Data Processing
h5py==3.8.0
pyarrow==12.0.0
tables==3.8.0

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1

# Logging and Experiment Tracking
tensorboard==2.13.0
wandb==0.15.3
mlflow==2.3.2

# Configuration
pyyaml==6.0
omegaconf==2.3.0
hydra-core==1.3.2

# Progress Bars
tqdm==4.65.0

# Privacy and Security
opacus==1.4.0  # Differential Privacy
cryptography==41.0.1

# Testing
pytest==7.3.1
pytest-cov==4.1.0
pytest-mock==3.10.0

# Code Quality
black==23.3.0
flake8==6.0.0
mypy==1.3.0
isort==5.12.0

# Documentation
sphinx==7.0.1
sphinx-rtd-theme==1.2.0

# Utilities
click==8.1.3
joblib==1.2.0
dill==0.3.6

# Data Download
gdown==4.7.1
requests==2.31.0
beautifulsoup4==4.12.2

# Performance
numba==0.57.0

# Type Checking
typing-extensions==4.6.2

# Jupyter (for analysis notebooks)
jupyter==1.0.0
ipywidgets==8.0.6

# Distributed Training
torch-distributed==0.1.0

# Additional utilities
psutil==5.9.5
gputil==1.4.0
py-cpuinfo==9.0.0
